requirements.txt
accelerate==1.12.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bitsandbytes==0.49.1
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
faiss-cpu==1.13.2
fastapi==0.129.0
filelock==3.20.0
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
packaging==26.0
pillow==12.0.0
psutil==7.2.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.1.15
rich==14.3.2
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
shellingham==1.5.4
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.3
transformers==5.1.0
typer==0.23.1
typer-slim==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
